move metrics to separate file
